chert js

+++ extractor/api/model/service.py
@@ -28,8 +28,8 @@ def encode_column(column_name, dataset):
 
 def add_parallel_processes(data):
     
-    data['duration'] = data['duration'].astype(np.int64)
-    data['starttime'] = data['starttime'].astype(np.int64)
+    #data['duration'] = data['duration'].astype(np.int64)
+    #data['starttime'] = data['starttime'].astype(np.int64)
 
     # Создание столбцов start и end
     data['start'] = data['starttime']
